Create README.md
Hey there!

So, I've created a repository that contains all the code and queries from my portfolio projects. The idea behind it is to share my work with others who might be interested in taking a look or even using it for their own projects.

Feel free to explore the repository, use the code and queries, and make them your own. I'd love to see what insights you come up with! I hope that this can be a helpful resource for you in your own work.

If you have any questions or feedback, please don't hesitate to let me know. I'm always happy to chat and help out however I can.

Thanks for checking out my work, and happy coding! :)